chore: move adf files to /lib/adf
