+++ src/modals/view-issue-modal.tsx
@@ -9,8 +9,8 @@ import { prefetchUsers } from "../api/get-users.query.js";
 import { modalsAtom, openModal } from "../atoms/modals.atom.js";
 import { env } from "../env.js";
 import { priorityMap } from "../issue.js";
-import { ADFRenderer } from "../lib/adf-renderer.js";
-import type { TopLevelNode } from "../lib/nodes.js";
+import { ADFRenderer } from "../lib/adf/adf-renderer.js";
+import type { TopLevelNode } from "../lib/adf/nodes.js";
 import { openIssueInBrowser } from "../lib/utils/openIssueInBrowser.js";
 import { PaddedText } from "../padded-text.js";
 import { useStdoutDimensions } from "../useStdoutDimensions.js";
