+++ src/padded-text.tsx
@@ -1,5 +1,5 @@
 import { Text } from "ink";
-import { ansiRegex } from "./lib/adf-renderer.js";
+import { ansiRegex } from "./lib/adf/adf-renderer.js";
 
 export const PaddedText = ({
   text,
